refactor: custom lotto를 만드는 메서드를 manualLottoGeneratorTest로 옮김
